+++ src/jp/ac/kobe_u/cs/sugar/csp/ArithmeticLiteral.java
@@ -6,5 +6,5 @@ import jp.ac.kobe_u.cs.sugar.SugarException;
 
 public abstract class ArithmeticLiteral implements Literal {
 	public abstract Set<IntegerVariable> getVariables();
-	public abstract int[] getBound(IntegerVariable v) throws SugarException;
+	public abstract long[] getBound(IntegerVariable v) throws SugarException;
 }
