+++ src/jp/ac/kobe_u/cs/sugar/csp/IntegerVariable.java
@@ -13,15 +13,15 @@ import jp.ac.kobe_u.cs.sugar.SugarException;
 public class IntegerVariable implements Comparable<IntegerVariable> {
 	private static final String AUX_PRE = "_$I";
 	private static String AUX_NAME_PREFIX = AUX_PRE;
-	private static int auxIntegerVariablesSize = 0;
+	private static long auxIntegerVariablesSize = 0;
 	private String name;
 	private IntegerDomain domain;
 	private boolean aux;
 	private String comment = null;
 	private boolean modified = true;
-	private int code;
-	private int value;
-	private int offset;
+	private long code;
+	private long value;
+	private long offset;
 	private boolean isDigit_;
 	private IntegerVariable[] vs = null;
 
@@ -29,7 +29,7 @@ public class IntegerVariable implements Comparable<IntegerVariable> {
 		AUX_NAME_PREFIX = AUX_PRE + pre;
 	}
 
-	public static void setIndex(int index) {
+	public static void setIndex(long index) {
 		auxIntegerVariablesSize = 0;
 	}
 
@@ -80,11 +80,11 @@ public class IntegerVariable implements Comparable<IntegerVariable> {
 		this.aux = aux;
 	}
 
-	public int getOffset() {
+	public long getOffset() {
 		return offset;
 	}
 
-	public void setOffset(int off) {
+	public void setOffset(long off) {
 		offset = off;
 	}
 
@@ -118,7 +118,7 @@ public class IntegerVariable implements Comparable<IntegerVariable> {
 		this.modified = modified;
 	}
 
-	public int bound(int lb, int ub) throws SugarException {
+	public long bound(long lb, long ub) throws SugarException {
 		IntegerDomain oldDomain = domain;
 		domain = domain.bound(lb, ub);
 		if (! domain.equals(oldDomain)) {
@@ -131,7 +131,7 @@ public class IntegerVariable implements Comparable<IntegerVariable> {
 	 * Returns the code value in the encoded representation. 
 	 * @return the code value in the encoded representation
 	 */
-	public int getCode() {
+	public long getCode() {
 		return code;
 	}
 
@@ -139,7 +139,7 @@ public class IntegerVariable implements Comparable<IntegerVariable> {
 	 * Sets the code value in the encoded representation. 
 	 * @param code the code value
 	 */
-	public void setCode(int code) {
+	public void setCode(long code) {
 		this.code = code;
 	}
 
@@ -147,7 +147,7 @@ public class IntegerVariable implements Comparable<IntegerVariable> {
 	 * Returns the value of the integer variable.
 	 * @return the value
 	 */
-	public int getValue() {
+	public long getValue() {
 		return value;
 	}
 
@@ -155,7 +155,7 @@ public class IntegerVariable implements Comparable<IntegerVariable> {
 	 * Sets the value of the integer variable.
 	 * @param value the value to set
 	 */
-	public void setValue(int value) {
+	public void setValue(long value) {
 		this.value = value;
 	}
 
@@ -184,8 +184,8 @@ public class IntegerVariable implements Comparable<IntegerVariable> {
 	}
 
 	public List<IntegerVariable> splitToDigits(CSP csp) throws SugarException {
-		int ub = domain.getUpperBound();
-		int b = csp.getBases()[0];
+		long ub = domain.getUpperBound();
+		long b = csp.getBases()[0];
 		int m = (int)Math.ceil(Math.log(ub+1)/Math.log(b));
 
 		vs = new IntegerVariable[m];
@@ -194,7 +194,7 @@ public class IntegerVariable implements Comparable<IntegerVariable> {
 		} else {
 			for (int i=0; i<m; i++, ub /= b) {
 				assert ub > 0;
-				int ubi = (i == m-1) ? ub : b-1;
+				long ubi = (i == m-1) ? ub : b-1;
 				IntegerDomain dom = new IntegerDomain(0, ubi);
 				vs[i] = new IntegerVariable(dom);
 				vs[i].isDigit(true);
